Add CanvasElementProperties
